ci: make every lane pass — 319 lint findings fixed, linter pinned, CRIU from PPA
CI has never been green. The check lane failed on golangci-lint findings
(319 uncapped: errcheck, noctx, gocritic evalOrder/unnamedResult/
paramTypeCombine, misspell, staticcheck), the e2e and stress lanes could
not install CRIU (ubuntu-24.04 dropped the package), and the web lane
failed lint.

- fix every lint finding at its root: real-context threading for servers
  and request paths (Listen/Dial/exec/NewRequest with ctx), explicit
  error sinks for writes to terminals, named results where gocritic
  demands them (with the local-declaration collisions cleaned up),
  evaluation-order hoists, switch conversions, param combines; frozen
  wire vocabulary ("cancelled", "CANCELLED") keeps its spelling under a
  nolint with the reason, free text is corrected
- transfer client: close each retry attempt's response body at once
  instead of deferring in the loop (four attempts held four connections)
- completion: drop a no-op strings.ReplaceAll(space, space) that gocritic
  flagged as dupArg
- pin golangci-lint-action to v2.13.2 (the config's v2 schema and an
  unpinned "latest" drifted apart); install CRIU from ppa:criu/ppa on
  the e2e and stress lanes since noble has no criu package
- web: derive the migrate dialog's effective mode instead of resetting
  state in an effect, wire the onMigrated callback, escape an apostrophe
- remove dead code the linter surfaced (drainAndClose, unused vars)

Verified locally: go build/vet/gofmt clean, golangci-lint 0 issues in
both capped and uncapped modes, go test ./... and go test -race
./internal/... green, web lint+build green, desktop frontend build,
cargo test and cargo build green.

+++ internal/filesystem/index.go
@@ -67,7 +67,7 @@ func IndexTree(root string, exclusions []string) (TreeIndex, error) {
 			stamp.Size = stat.Size
 			stamp.MtimeSec, stamp.MtimeNS = stat.Mtim.Sec, stat.Mtim.Nsec
 			stamp.Inode = stat.Ino
-			stamp.Device = uint64(stat.Dev)
+			stamp.Device = stat.Dev
 		} else {
 			stamp.Size = info.Size()
 		}
